// src/Utils/enums/auth.enum.ts
export enum GenderEnum {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum RoleEnum {
  ADMIN = "ADMIN",
  USER = "USER",
}

export enum SignatureEnum{
  USER = 0,
  ADMIN = 1
}

export enum TokinTypeEnum {
  ACCESS = "ACCESS",
  REFRESH = "REFRESH"
}

export enum LogoutTypeEnum {
  LOGOUT = "LOGOUT",
  LOGOUT_FROM_ALL = "LOGOUT_FROM_ALL"
}

export enum ProviderEnum{
  GOOGLE = "GOOGLE",
  SYSTEM = "SYSTEM"
}